More Mod-Tap and minor switch of keys in raise layer
Mod-tap change:
---------------

Put KC_ENT behind mod taps. Introduce an extra enter so that I still
can push it if I happen to hold the modifier hiding the enter.

Reason for the mod-tap, something I've really tried to avoid, is that
on Mac it's quite uncomfortable for me to only have the GUI-key on
left pinky.

Minor switch of keys in raise layer:
------------------------------------

Put {[ in a position that mimics that on regular keyboards. Previously
they were over and under.

Instead of this:

```
[]
{}
```

Now they're like this:

```
{[]}
```

The reason to put them on the lower row is to have them on the same
row as regular parentheses in the lower layer. Also mimicking the
alt-gr modification on regular keyboards, so that () on lower layer maps to [] on
the raise layer.

Side consequence of this is that the "shift" thinking for ; and =
doesn't work any longer. Those keys together with their "shift" state,
: and + are now found next to each other in the _Raise layer.